+++ src/plugins/liveSearch/components.ts
@@ -1,2 +1,3 @@
 export * from './no/noLiveSearch.component';
 export * from './basic/basicLiveSearch.component';
+export * from './edit/editLiveSearch.component';