+++ src/plugins/liveSearch/index.ts
@@ -2,3 +2,4 @@ export * from './liveSearch.interface';
 
 export * from './basic/basicLiveSearch.interface';
 export * from './no/noLiveSearch.interface';
+export * from './edit/editLiveSearch.interface';