+++ src/plugins/normalState/components.ts
@@ -1,1 +1,2 @@
 export * from './basic/basicNormalState.component';
+export * from './edit/editNormalState.component';