+++ src/plugins/normalState/index.ts
@@ -1,3 +1,4 @@
 export * from './normalState.interface';
 
 export * from './basic/basicNormalState.interface';
+export * from './edit/editNormalState.interface';